+++ src/main/java/si/matjazcerkvenik/eventlogger/parsers/HttpPostParser.java
@@ -70,6 +70,7 @@ public class HttpPostParser implements IEventParser {
             e.setHost(dRequest.getRemoteHost());
             e.setEndpoint(dRequest.getRequestUri());
             e.setEventSource(dRequest.getRemoteHost());
+            e.setLogfile("unknown");
             e.setIdent("http.post.plain-text");
             if (dRequest.getParameterMap().containsKey("ident")) {
                 e.setIdent(dRequest.getParameterMap().get("ident"));
@@ -77,6 +78,9 @@ public class HttpPostParser implements IEventParser {
             e.setPid("0");
             if (dRequest.getParameterMap().containsKey("pid")) {
                 e.setPid(dRequest.getParameterMap().get("pid"));
+            }
+            if (dRequest.getParameterMap().containsKey("file")) {
+                e.setLogfile(dRequest.getParameterMap().get("file"));
             }
             if (dRequest.getParameterMap().containsKey("tag")) {
                 e.setTag(dRequest.getParameterMap().get("tag"));