+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Now all events are received on `/event/webhook/<name>` endpoint.
 * [FEATURE] Event rules view
 * [ENHANCEMENT] Show ajax status when loading events from DB
 * [ENHANCEMENT] Added `Home` icon in toolbar
+* [ENHANCEMENT] Added new parameter to event: `logfile` - originating log file.
 * [FIX] Corrections for raising and clearing alarms